refactor(internal): reorganize packages and improve git staging
