+++ internal/routing/model_selector.go
