+++ internal/routing/model_selector_test.go
